server/requirements.txt
pillow
fastapi[standard]
uvicorn
